Moved `_Labels` to `_encoding.EncodedLabels`. added str repr
